% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_state.R
\name{add_secondary_union_fields}
\alias{add_secondary_union_fields}
\title{Add secondary union fields to state}
\usage{
add_secondary_union_fields(state, library)
}
\arguments{
\item{state}{Parcel MCMC state}

\item{library}{Secondary library}
}
\value{
State with union fields populated
}
\description{
Add secondary union fields to state
}
\keyword{internal}
