% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{aggregate_move_stats}
\alias{aggregate_move_stats}
\title{Aggregate move statistics across chains}
\usage{
aggregate_move_stats(valid_chains)
}
\arguments{
\item{valid_chains}{List of valid chain results}
}
\value{
Combined stats data.table
}
\description{
Aggregate move statistics across chains
}
