% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{compute_ess_parcel}
\alias{compute_ess_parcel}
\title{Compute effective sample size from trajectory}
\usage{
compute_ess_parcel(x, lag_max = 50)
}
\arguments{
\item{x}{Numeric vector}

\item{lag_max}{Maximum lag for autocorrelation}
}
\value{
ESS estimate
}
\description{
Compute effective sample size from trajectory
}
