% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{count_secondary_turnover}
\alias{count_secondary_turnover}
\title{Count secondary birth/death events}
\usage{
count_secondary_turnover(parcel_samples)
}
\arguments{
\item{parcel_samples}{List of parcel states}
}
\value{
List with births, deaths, net_change
}
\description{
Count secondary birth/death events
}
