% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{count_unique_parcel_states}
\alias{count_unique_parcel_states}
\title{Count unique parcel states}
\usage{
count_unique_parcel_states(parcel_samples)
}
\arguments{
\item{parcel_samples}{List of parcel states}
}
\value{
Integer count
}
\description{
Two states are equal if they have the same X (set of parcels).
}
