% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_state.R
\name{create_lcc_only_state}
\alias{create_lcc_only_state}
\title{Create initial state from LCC-only (no secondaries)}
\usage{
create_lcc_only_state(lcc_parcels, library, parcel_graph)
}
\arguments{
\item{lcc_parcels}{Character vector of parcel IDs for the LCC}

\item{library}{Secondary library}

\item{parcel_graph}{igraph object}
}
\value{
Parcel MCMC state
}
\description{
Convenience function for creating a single-component state.
}
