% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_state.R
\name{get_lcc_capacity}
\alias{get_lcc_capacity}
\title{Get capacity of LCC}
\usage{
get_lcc_capacity(state, library, parcel_graph)
}
\arguments{
\item{state}{Parcel MCMC state}

\item{library}{Secondary library}

\item{parcel_graph}{igraph object}
}
\value{
Integer capacity
}
\description{
Get capacity of LCC
}
