% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_state.R
\name{get_lcc_parcels}
\alias{get_lcc_parcels}
\title{Get parcels in the LCC}
\usage{
get_lcc_parcels(state, library)
}
\arguments{
\item{state}{Parcel MCMC state}

\item{library}{Secondary library}
}
\value{
Character vector of parcel IDs in LCC
}
\description{
LCC = X minus all secondary parcels
}
