% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{get_parcel_set_neighbors}
\alias{get_parcel_set_neighbors}
\title{Compute neighbors of a set of parcels}
\usage{
get_parcel_set_neighbors(parcel_ids, parcel_graph, neighbor_cache = NULL)
}
\arguments{
\item{parcel_ids}{Character vector of parcel IDs}

\item{parcel_graph}{igraph object}

\item{neighbor_cache}{Optional named list of precomputed neighbors (for speed)}
}
\value{
Character vector of neighbor parcel IDs (excluding input parcels)
}
\description{
Compute neighbors of a set of parcels
}
