% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_state.R
\name{get_secondary_parcels}
\alias{get_secondary_parcels}
\title{Get all parcels belonging to secondary components}
\usage{
get_secondary_parcels(state, library)
}
\arguments{
\item{state}{Parcel MCMC state}

\item{library}{Secondary library}
}
\value{
Character vector of parcel IDs in secondaries
}
\description{
Get all parcels belonging to secondary components
}
