% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{get_sticky_parcels}
\alias{get_sticky_parcels}
\title{Get summary of sticky parcel-units}
\usage{
get_sticky_parcels(stickiness_dt)
}
\arguments{
\item{stickiness_dt}{data.table from compute_parcel_stickiness()}
}
\value{
list with always_in and rarely_in parcel IDs and counts
}
\description{
Get summary of sticky parcel-units
}
