% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{hydrate_libraries}
\alias{hydrate_libraries}
\title{Hydrate all block libraries}
\usage{
hydrate_libraries(libraries)
}
\arguments{
\item{libraries}{List containing secondary_library and lcc_library}
}
\value{
Hydrated libraries
}
\description{
Wrapper to hydrate both secondary and LCC libraries.
}
