% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{hydrate_library}
\alias{hydrate_library}
\title{Hydrate library}
\usage{
hydrate_library(library)
}
\arguments{
\item{library}{Block library object}
}
\value{
Hydrated library with hash index
}
\description{
Ensures libraries are in the compact representation:
\itemize{
\item blocks stored as integer indices into parcel_names
\item neighbor_indices stored as integer indices
\item block_hashes rebuilt when missing
}
}
