% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_state.R
\name{initialize_parcel_state}
\alias{initialize_parcel_state}
\title{Initialize parcel MCMC state}
\usage{
initialize_parcel_state(parcel_ids, secondary_block_ids, library, parcel_graph)
}
\arguments{
\item{parcel_ids}{Character vector of all parcel IDs in the overlay}

\item{secondary_block_ids}{Integer vector of secondary block IDs (from L_sec)}

\item{library}{Secondary library}

\item{parcel_graph}{igraph object}
}
\value{
Parcel MCMC state
}
\description{
Creates a state from a set of parcel IDs. Secondaries must be specified
as block IDs (not derived geometrically).
}
