% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{library_block_indices}
\alias{library_block_indices}
\title{Get block indices from library}
\usage{
library_block_indices(library, block_id)
}
\arguments{
\item{library}{Block library}

\item{block_id}{Integer block ID}
}
\value{
Integer vector of parcel indices
}
\description{
Get block indices from library
}
