% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{library_block_parcels}
\alias{library_block_parcels}
\title{Get block parcel IDs from library}
\usage{
library_block_parcels(library, block_id)
}
\arguments{
\item{library}{Block library}

\item{block_id}{Integer block ID}
}
\value{
Character vector of parcel IDs
}
\description{
Get block parcel IDs from library
}
