% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{library_blocks_parcels}
\alias{library_blocks_parcels}
\title{Get parcel IDs for multiple block IDs}
\usage{
library_blocks_parcels(library, block_ids)
}
\arguments{
\item{library}{Block library}

\item{block_ids}{Integer vector of block IDs}
}
\value{
Character vector of parcel IDs
}
\description{
Get parcel IDs for multiple block IDs
}
