% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{parcel_ids_to_indices}
\alias{parcel_ids_to_indices}
\title{Convert parcel IDs to integer indices}
\usage{
parcel_ids_to_indices(parcel_ids, parcel_names)
}
\arguments{
\item{parcel_ids}{Character or integer vector of parcel IDs}

\item{parcel_names}{Character vector of all parcel names}
}
\value{
Integer vector of indices
}
\description{
Convert parcel IDs to integer indices
}
