% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbta_map.R
\name{print.mbta_map}
\alias{print.mbta_map}
\title{Print Method for mbta_map}
\usage{
\method{print}{mbta_map}(x, ...)
}
\arguments{
\item{x}{An mbta_map object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Display formatted summary of mbta_map object
}
