% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_particle.R
\name{print.ParticleState}
\alias{print.ParticleState}
\title{Print Method for ParticleState}
\usage{
\method{print}{ParticleState}(x, ...)
}
\arguments{
\item{x}{A ParticleState object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Display formatted summary of particle state
}
