% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_feasibility.R
\name{summarize_parcel_feasibility}
\alias{summarize_parcel_feasibility}
\title{Summarize feasibility analysis for reporting}
\usage{
summarize_parcel_feasibility(feasibility_result, parcel_graph)
}
\arguments{
\item{feasibility_result}{Output from analyze_parcel_feasibility()}

\item{parcel_graph}{igraph object}
}
\value{
List with summary statistics for Quarto report
}
\description{
Summarize feasibility analysis for reporting
}
