% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbta_map.R
\name{summary.mbta_map}
\alias{summary.mbta_map}
\title{Summary Method for mbta_map}
\usage{
\method{summary}{mbta_map}(object, ...)
}
\arguments{
\item{object}{An mbta_map object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns object
}
\description{
Display extended statistics for mbta_map object
}
