% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_mcmc_kernels.R
\name{symmetric_birth_death_move}
\alias{symmetric_birth_death_move}
\title{Symmetric Birth/Death Move}
\usage{
symmetric_birth_death_move(
  state,
  library,
  parcel_graph,
  constraints,
  capacity_prior_lambda,
  k_prior_lambda,
  birth_tilt_lambda,
  neighbor_idx = NULL
)
}
\arguments{
\item{state}{Current state}

\item{library}{Secondary library}

\item{parcel_graph}{igraph object}

\item{constraints}{MBTA constraints}

\item{neighbor_idx}{Precomputed neighbor indices}
}
\value{
List with new_state, accepted, direction, details
}
\description{
Proposes birth or death by sampling uniformly from universe U = addable U removable.
}
\details{
Proposal q(y|x):
\enumerate{
\item Build U = {addable blocks} U {removable blocks}
\item Sample block B uniformly from U
\item If B is addable, propose birth; if removable, propose death
}

MH Ratio:
alpha = (pi(y)/pi(x)) * (q(x|y)/q(y|x))
q(x|y)/q(y|x) = |U_old| / |U_new| = Z_old / Z_new
}
