% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{trace_centroids}
\alias{trace_centroids}
\title{Compute centroid trajectory}
\usage{
trace_centroids(parcel_samples, parcel_graph)
}
\arguments{
\item{parcel_samples}{List of parcel states}

\item{parcel_graph}{igraph with centroid_x, centroid_y attributes}
}
\value{
data.table with step, centroid_x, centroid_y
}
\description{
Compute centroid trajectory
}
