% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{trace_geographic_spread}
\alias{trace_geographic_spread}
\title{Compute geographic spread (bounding box area) over time}
\usage{
trace_geographic_spread(parcel_samples, parcel_graph)
}
\arguments{
\item{parcel_samples}{List of parcel states}

\item{parcel_graph}{igraph with centroid_x, centroid_y}
}
\value{
Numeric vector of bounding box areas
}
\description{
Compute geographic spread (bounding box area) over time
}
