% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_metrics.R
\name{trace_n_secondaries}
\alias{trace_n_secondaries}
\title{Extract secondary component count trajectory}
\usage{
trace_n_secondaries(parcel_samples)
}
\arguments{
\item{parcel_samples}{List of parcel states}
}
\value{
Integer vector
}
\description{
Extract secondary component count trajectory
}
