% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_config.R
\name{define_constraints}
\alias{define_constraints}
\title{Define constraints from district requirements}
\usage{
define_constraints(district_data)
}
\arguments{
\item{district_data}{List containing district_requirements from data loading}
}
\value{
List of constraints for MCMC
}
\description{
Define constraints from district requirements
}
