% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_config.R
\name{define_constraints}
\alias{define_constraints}
\title{Define constraints from Norwood requirements}
\usage{
define_constraints(norwood_data)
}
\arguments{
\item{norwood_data}{List containing norwood_requirements from data loading}
}
\value{
List of constraints for MCMC
}
\description{
Define constraints from Norwood requirements
}
