% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_data_loading.R
\name{get_district_paths}
\alias{get_district_paths}
\title{Helper function to get paths given a district name}
\usage{
get_district_paths(district_name, data_root = "data")
}
\arguments{
\item{district_name}{Name of district (e.g., "Norwood")}

\item{data_root}{Path to data directory (default: "data")}
}
\value{
Named list with paths
}
\description{
Helper function to get paths given a district name
}
