% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtazone-package.R
\docType{package}
\name{mbtazone-package}
\alias{mbtazone}
\alias{mbtazone-package}
\title{mbtazone: Implements the MBTA Communities Compliance Model in R}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdhidalgo/mbtazone}
  \item Report bugs at \url{https://github.com/fdhidalgo/mbtazone/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
