% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbtazone-package.R
\docType{package}
\name{mbtazone-package}
\alias{mbtazone}
\alias{mbtazone-package}
\title{mbtazone: Implements the MBTA Communities Compliance Model in R}
\description{
Implements the Massachusetts MBTA Communities Act compliance model in R, providing automated zoning compliance assessment tools. Replaces manual Excel-based workflows with robust calculation functions for housing unit capacity and interactive planning capabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdhidalgo/mbtazone}
  \item Report bugs at \url{https://github.com/fdhidalgo/mbtazone/issues}
}

}
\author{
\strong{Maintainer}: Fernando Hidalgo \email{dhidalgo@mit.edu}

}
\keyword{internal}
