% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_parcel_library.R
\name{select_maxmin_distant_seeds}
\alias{select_maxmin_distant_seeds}
\title{Select k maximally distant seed components using greedy max-min algorithm}
\usage{
select_maxmin_distant_seeds(viable_components, k)
}
\arguments{
\item{viable_components}{data.table from find_viable_station_components()}

\item{k}{Number of seeds to select}
}
\value{
Integer vector of row indices in viable_components
}
\description{
Select k maximally distant seed components using greedy max-min algorithm
}
